Multiple polygons/bugs fixes
