+++ config.cfg
@@ -2,7 +2,7 @@
 scanDir = /Users/jazminromero/development_github/development_fork/siglib/SigLib/scanDir
 tmpDir = /Users/jazminromero/development_github/development_fork/siglib/SigLib/tmpDir
 projDir = /Users/jazminromero/development_github/development_fork/siglib/SigLib/projDir
-vectDir = /Users/jazminromero/development_github/development/siglib/SigLib/roiDir
+vectDir = /Users/jazminromero/development_github/development/siglib/SigLib/roiDir/poly5
 imgDir = /Users/jazminromero/development_github/development_fork/siglib/SigLib/imgDir
 logDir = /Users/jazminromero/development_github/development_fork/siglib/SigLib/logDir
 outDir = /Users/jazminromero/Desktop/outDir
@@ -33,7 +33,7 @@ proj = lcc
 projSRID = 96718
 imgtypes = amp
 imgformat = GTiff
-roi = ArcticBay
+roi = ExperimentalFarm
 roiprojSRID = 4326
 mask =  
 crop =
